+++ lib/than_pkg.dart
@@ -1,7 +1,9 @@
 import 'dart:io';
 
+import 'package:than_pkg/enums/screen_orientation_types.dart';
 import 'package:than_pkg/than_pkg_android.dart';
 import 'package:than_pkg/than_pkg_interface.dart';
+import 'package:than_pkg/than_pkg_linux.dart';
 
 class ThanPkg implements ThanPkgInterface {
   static ThanPkg get platform => _createInstance();
@@ -9,6 +11,8 @@ class ThanPkg implements ThanPkgInterface {
   static ThanPkg _createInstance() {
     if (Platform.isAndroid) {
       return ThanPkgAndroid();
+    } else if (Platform.isLinux) {
+      return ThanPkgLinux();
     }
     return ThanPkg();
   }
@@ -37,12 +41,6 @@ class ThanPkg implements ThanPkgInterface {
   @override
   Future<String?> getWifiAddress() {
     // TODO: implement getWifiAddress
-    throw UnimplementedError();
-  }
-
-  @override
-  Future<List?> getWifiAddressList() {
-    // TODO: implement getWifiAddressList
     throw UnimplementedError();
   }
 
@@ -87,4 +85,39 @@ class ThanPkg implements ThanPkgInterface {
     // TODO: implement checkAndRequestPackageInstallPermission
     throw UnimplementedError();
   }
+
+  @override
+  Future<List<String>> getWifiAddressList() {
+    // TODO: implement getWifiAddressList
+    throw UnimplementedError();
+  }
+  
+  @override
+  Future<void> genVideoCover({required Comparable<String> outDirPath, required List<String> videoPathList, int iconSize = 300}) {
+    // TODO: implement getVideoCover
+    throw UnimplementedError();
+  }
+  
+  @override
+  Future<Map<String, dynamic>> getAndroidDeviceInfo() {
+    // TODO: implement getAndroidDeviceInfo
+    throw UnimplementedError();
+  }
+
+  
+  @override
+  Future<void> requestScreenOrientation({required ScreenOrientationTypes type, bool reverse = false}) {
+    // TODO: implement requestScreenOrientation
+    throw UnimplementedError();
+  }
+  
+  @override
+  Future<ScreenOrientationTypes?> checkScreenOrientation() {
+    // TODO: implement checkScreenOrientation
+    throw UnimplementedError();
+  }
+
+
+ 
+
 }