+++ lib/than_pkg_interface.dart
@@ -1,8 +1,7 @@
+import 'package:than_pkg/enums/screen_orientation_types.dart';
+
 abstract class ThanPkgInterface {
-  Future<String?> getPlatformVersion();
-  Future<String?> getLocalIpAddress();
-  Future<String?> getWifiAddress();
-  Future<List<dynamic>?> getWifiAddressList();
+  Future<List<String>> getWifiAddressList();
   Future<bool> openUrl({required String url});
   Future<void> toggleFullScreen({required bool isFullScreen});
   Future<void> genPdfCover({
@@ -10,9 +9,25 @@ abstract class ThanPkgInterface {
     required List<String> pdfPathList,
     int iconSize = 300,
   });
-  Future<void> toggleKeepScreen({required bool isKeep});
+
+  Future<void> genVideoCover({
+    required String outDirPath,
+    required List<String> videoPathList,
+    int iconSize = 300,
+  });
+  //android only
+  Future<ScreenOrientationTypes?> checkScreenOrientation();
+  Future<void> requestScreenOrientation({
+    required ScreenOrientationTypes type,
+    bool reverse = false,
+  });
+  Future<Map<String, dynamic>> getAndroidDeviceInfo();
   Future<bool> isStoragePermissionGranted();
   Future<void> requestStoragePermission();
   Future<void> checkAndRequestPackageInstallPermission();
   Future<String?> getDeviceId();
+  Future<String?> getPlatformVersion();
+  Future<String?> getLocalIpAddress();
+  Future<String?> getWifiAddress();
+  Future<void> toggleKeepScreen({required bool isKeep});
 }